pyproject.toml
[tool.poetry]
name = "graphrag"
version = "0.1.0"
description = ""
authors = ["stefano <sturni93@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
pymupdf = ">=1.26.3,<2.0.0"
langdetect = ">=1.0.9,<2.0.0"
python-dotenv = ">=1.1.1,<2.0.0"
langchain = ">=0.3.0,<0.4.0"
langchain-openai = ">=0.2.0,<0.3.0"
langchain-community = ">=0.3.0,<0.4.0"
langgraph = ">=0.2.0,<0.3.0"
langchain-neo4j = "^0.5.0"
tavily-python = ">=0.5.0,<1.0.0"

[tool.poetry.group.test.dependencies]
pytest = ">=7.0.0"
pytest-mock = ">=3.10.0"
pytest-cov = ">=4.0.0"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:.*SwigPyPacked.*:DeprecationWarning",
    "ignore:.*SwigPyObject.*:DeprecationWarning",
    "ignore:.*swigvarlink.*:DeprecationWarning"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
